pyproject.toml
[tool.poetry]
name = "rag_search_retrieval"
version = "0.1.0"
description = "Experiment with search and retrieval approaches for an RAG application."
authors = ["Fuzzy Labs <info@fuzzylabs.ai>"]
license = "Apache Software License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10, <3.13"
pymilvus = {extras = ["model"], version = "^2.4.4"}
langchain-community = "^0.2.10"
pypdf = "^4.3.1"
langchain-text-splitters = "^0.2.2"
jupyter = "^1.0.0"
sentence-transformers = "^3.0.1"
scipy = "1.13.0"
flagembedding = "^1.2.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
licensecheck = "^2024.1.2"
pandas-stubs = "^2.2.2.240603"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=rag_search_retrieval --cov-report term-missing"
testpaths = [
    "tests"
]

# mypy configuration
[tool.mypy]
show_error_codes = true
exclude = ["docs", "tests", "LICENSE", "**/*.ipynb"]
strict = true
namespace_packages = true
ignore_missing_imports = true
disallow_any_generics = false
warn_return_any = false

# black configuration
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
target-version = "py310"

# Match black. Note that this also checks comment line length, but black does not format comments.
line-length = 88

show-fixes = true

[tool.ruff.lint]
ignore-init-module-imports = true
select = [
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "Q",    # flake8-quotes
    "ISC",  # flake8-implicit-str-concat
    "F",    # pyflakes
    "D",    # pydocstyle
    "E",    # pycodestyle error
    "W",    # pycodestyle warning
    "N",    # pep8-naming
    "I",    # isort
    "PL",   # pylint rules from categories "Convention", "Error", and "Warning"
    "PLE",  # ruff currently implements only a subset of pylint's rules
    "PLW",  # pylint warning
    "PLR",  # pylint refactor
    "UP",   # pyupgrade
    "C",    # Complexity (mccabe+) & comprehensions
]
ignore = [
    "UP006", # See https://github.com/bokeh/bokeh/issues/13143
    "UP007", # See https://github.com/bokeh/bokeh/pull/13144
]

[tool.ruff.lint.pylint]
# Maximum number of arguments allowed for a function or method definition.
max-args = 6

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 5.
max-complexity = 5


# typos configuration
[tool.typos.files]
extend-exclude=[".gitignore", "LICENSE", ".*", "**/*.ipynb"]

[tool.typos.default.extend-words]
center = "center"

[tool.typos.default]
locale="en-gb"

# Bandit configuration
[tool.bandit]
exclude_dirs = []

[tool.bandit.assert_used]
skips = ['*test.py', '*/test_*.py']
